Removed dash in package name and moved tests to inside of package
