+++ exoticism/tests/.pytest_cache/v/cache/nodeids
@@ -1,0 +1,8 @@
+[
+  "test_configfile.py::test_main_sections",
+  "test_configfile.py::test_data_paths",
+  "test_configfile.py::test_setup",
+  "test_configfile.py::test_smooth_model",
+  "test_configfile.py::test_constants",
+  "test_limb_darkening.py::test_limb_dark_fit"
+]
